pyproject.toml
[tool.poetry]
name = "fastapi-boilerplate"
version = "1.0.0"
description = "Production-ready FastAPI backend boilerplate"
authors = ["Parth Joshi <parthjoshi.1618@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
gunicorn = "^21.2.0"
sqlalchemy = "^2.0.0"
asyncpg = "^0.29.0"
alembic = "^1.13.0"
pyjwt = "^2.8.0"
passlib = { extras = ["argon2"], version = "^1.7.4" }
argon2-cffi = "^23.1.0"
pydantic = { extras = ["email"], version = "^2.7.0" }
pydantic-settings = "^2.2.0"
python-multipart = "^0.0.9"
structlog = "^24.1.0"
loguru = "^0.7.2"
slowapi = "^0.1.9"
redis = { extras = ["hiredis"], version = "^5.0.0" }
python-dotenv = "^1.0.0"
aiosqlite = "^0.21.0"
opentelemetry-api = "^1.21.0"
opentelemetry-sdk = "^1.21.0"
opentelemetry-instrumentation-fastapi = "^0.43b0"
opentelemetry-instrumentation-sqlalchemy = "^0.43b0"
opentelemetry-instrumentation-redis = "^0.43b0"
opentelemetry-exporter-otlp = "^1.21.0"
opentelemetry-exporter-jaeger = "^1.21.0"
aiocache = "^0.12.2"
tenacity = "^8.2.3"
circuitbreaker = "^2.0.0"
celery = { extras = ["redis"], version = "^5.3.0" }
prometheus-client = "^0.20.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.27.0"
ruff = "^0.3.0"
mypy = "^1.9.0"
pre-commit = "^3.7.0"
types-passlib = "^1.7.7"
types-redis = "^4.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "N",
  "W",
  "UP",
  "B",
  "A",
  "C4",
  "DTZ",
  "T10",
  "EM",
  "ISC",
  "ICN",
  "PIE",
  "PT",
  "Q",
  "RSE",
  "RET",
  "SIM",
  "TID",
  "ERA",
  "PL",
  "RUF",
]
ignore = [
  "E501",    # Line too long (handled by formatter)
  "PLR0913", # Too many arguments
  "PLR2004", # Magic value
  "EM101",
  "ISC001",  # Fixes the formatter conflict warning
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
strict_equality = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["slowapi.*", "loguru.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
